Fix dmenu forking
If the scripted dmenu is in $PATH (for example ~/bin/dmenu), the script
calls itself recursively causing the system to run out of memory and go
to a grinding hault.

The xmonad.hs takes the executable as a variable and then executes it,
if the script would still have $() it would try to run the return code
(0). With these modifications the haskell file can execute the
executable returned by dmenu